requirements.txt
django==1.8
google-api-python-client
tweepy
python-social-auth==0.2.21
oauth2client==1.3
django-allauth
urllib3
python-gcm
django-star-rating
